+++ GeoIP2/Responses/CountryResponse.cs
@@ -14,8 +14,14 @@ namespace MaxMind.GeoIP2.Responses
     /// </summary>
     public class CountryResponse : AbstractCountryResponse
     {
+        /// <summary>
+        /// Constructor
+        /// </summary>
         public CountryResponse() { }
 
+        /// <summary>
+        /// Constructor
+        /// </summary>
         public CountryResponse(Continent continent = null, Country country = null, Model.MaxMind maxMind = null,
             Country registeredCountry = null, RepresentedCountry representedCountry = null, Traits traits = null)
             : base(continent, country, maxMind, registeredCountry, representedCountry, traits)